For rebuilding testing packages
