+++ api/src/main/java/org/apache/cloudstack/diagnostics/DiagnosticsType.java
@@ -16,16 +16,27 @@
 // specific language governing permissions and limitations
 // under the License.
 //
-package org.apache.cloudstack.diangostics;
+package org.apache.cloudstack.diagnostics;
 
-import com.cloud.exception.AgentUnavailableException;
-import com.cloud.exception.InvalidParameterValueException;
-import org.apache.cloudstack.api.command.admin.diagnostics.ExecuteDiagnosticsCmd;
+public enum DiagnosticsType {
+    PING("ping"), TRACEROUTE("traceroute"), ARPING("arping");
 
-import java.util.Map;
+    private String value;
 
-public interface DiagnosticsService {
+    DiagnosticsType(String value) {
+        this.value = value;
+    }
 
-    Map<String, String> runDiagnosticsCommand(ExecuteDiagnosticsCmd cmd) throws AgentUnavailableException, InvalidParameterValueException;
+    public String getValue() {
+        return value;
+    }
 
+    public static DiagnosticsType getCommand(String cmd) {
+        for (DiagnosticsType type : DiagnosticsType.values()) {
+            if (type.value.equalsIgnoreCase(cmd)) {
+                return type;
+            }
+        }
+        return null;
+    }
 }
